Add `SimplePrettyApp` and `PrettyException`
This pull request proposes a `SimplePrettyApp` type with two related functions; like `SimpleApp` of the `RIO` package and its two related functions, but also supporting pretty printing functionality.

It also proposes a `PrettyException` type that represents pretty exceptions. That is, exceptions that are instances of `Pretty`. It includes Haddock documentation that illustrates how such pretty exceptions can be used.